Refactoring based on PR comments.
This moves standard mode to its own driver. The channel structs are
no longer separate from the driver; instead, the driver implements
I2sRxChannel and/or I2sTxChannel depending on a generic it takes,
which can be I2sRxSupported and/or I2sTxSupported.

Alas, the separate I2sStdDriverInternal struct is still there. This is
required because it has to be pinned in memory to serve as callback
data for the interrupt handler, while I2sStdDriver can be moved
around as Rust pleases.

Stubs configs have been created for PCM and PDM modes.